Strain Energy Density now expressed simply as I1.
Previously, we approximated strain energy density in the skin using
a Neo-Hookean material.  Now, we drop the Neo-Hookean approximation
and simply say we're reporting the First Invariant of the Left
Cauchy-Green deformation tensor (which is the sum of the squares of
the principal stretches).

Thanks to Dr Tom Pence from Michigan State University for advice
on this topic.